+++ app/modules/cdr/utils_cdr.py
@@ -2,6 +2,7 @@ import logging
 from datetime import UTC, datetime
 from uuid import UUID, uuid4
 
+import pandas as pd
 from fastapi import (
     HTTPException,
 )
@@ -154,3 +155,101 @@ async def check_request_consistency(
                 detail="Document is not related to this seller.",
             )
     return db_product
+
+
+def construct_dataframe_from_users_purchases(
+    users_purchases: dict[str, list[models_cdr.Purchase]],
+    users_answers: dict[str, list[models_cdr.CustomData]],
+    users: list[models_core.CoreUser],
+    products: list[models_cdr.CdrProduct],
+    variants: list[models_cdr.ProductVariant],
+    data_fields: dict[UUID, list[models_cdr.CustomDataField]],
+) -> pd.DataFrame:
+    """
+    Construct a pandas DataFrame from a dict of users purchases.
+
+    Args:
+        users_purchases (dict[str, list[models_cdr.Purchase]]): A dict of users purchases.
+
+    Returns:
+        pd.DataFrame: A pandas DataFrame.
+    """
+    columns = ["Nom", "Prénom", "Surnom", "Email"]
+    data = ["", "", "", "Prix"]
+    field_to_column = {}
+    variant_to_column = {}
+    for product in products:
+        product_variants = [
+            variant for variant in variants if variant.product_id == product.id
+        ]
+        columns.extend(
+            [f"{product.name_fr} : {variant.name_fr}" for variant in product_variants],
+        )
+        fields = data_fields.get(product.id, [])
+        columns.extend(
+            [f"{product.name_fr} : {field.name}" for field in fields],
+        )
+        data.extend([str(variant.price) for variant in product_variants])
+
+        data.extend([" " for _ in fields])
+        field_to_column.update(
+            {field.id: f"{product.name_fr} : {field.name}" for field in fields},
+        )
+        variant_to_column.update(
+            {
+                variant.id: f"{product.name_fr} : {variant.name_fr}"
+                for variant in product_variants
+            },
+        )
+    columns.append("Panier payé")
+    data.append(" ")
+    columns.append("Commentaire")
+    data.append(" ")
+    df = pd.DataFrame(
+        columns=columns,
+    )
+    df = pd.concat(
+        [
+            df,
+            pd.DataFrame(
+                data=[data],
+                columns=columns,
+            ),
+        ],
+    )
+    for user_id, purchases in users_purchases.items():
+        for purchase in purchases:
+            df.loc[
+                user_id,
+                variant_to_column[purchase.product_variant_id],
+            ] = purchase.quantity
+
+    for user_id, answers in users_answers.items():
+        for answer in answers:
+            column = field_to_column.get(answer.field_id)
+            if column:
+                df.loc[
+                    user_id,
+                    field_to_column[answer.field_id],
+                ] = answer.value
+
+    for user_id in df.index:
+        user = next((u for u in users if u.id == user_id), None)
+        if user is None:
+            continue
+        df.loc[user_id, "Nom"] = user.name
+        df.loc[user_id, "Prénom"] = user.firstname
+        df.loc[user_id, "Surnom"] = user.nickname
+        df.loc[user_id, "Email"] = user.email
+        if all(purchase.validated for purchase in users_purchases[user_id]):
+            df.loc[user_id, "Panier payé"] = True
+        else:
+            df.loc[user_id, "Panier payé"] = False
+            df.loc[user_id, "Commentaire"] = "Manquant : " + ", ".join(
+                variant_to_column[purchase.product_variant_id]
+                for purchase in users_purchases[user_id]
+                if not purchase.validated
+            )
+    for column in df.columns:
+        df[column] = df[column].apply(lambda x: x if pd.notna(x) else "")
+    return df
